requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
Pillow==10.0.1
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-storages==1.14.2
boto3==1.29.7
razorpay==1.3.0
python-decouple==3.8
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
django-extensions==3.2.3
requests==2.31.0
dj-database-url==2.1.0
